Route peer share cache writes through owner
